autonomous-coding/requirements.txt
# Dependencies

GitPython>=3.1.0
PyYAML>=6.0
